refactor scoping analysis
